Crate governance_canister

Structs

CanisterEnv

LedgerCanister

Constants

LOG_PREFIX

STABLE_MEM_BUFFER_SIZE
Size of the buffer for stable memory reads and writes.

Statics

GOVERNANCE

Functions

canister_heartbeat

canister_init

canister_init_

canister_post_upgrade

canister_pre_upgrade

claim_gtc_neurons

claim_gtc_neurons_

create_or_refresh_neuron

current_authz

current_authz_

execute_eligible_proposals
DEPRECATED: Proposals are now executed on every vote.

get_full_neuron
Returns the full neuron corresponding to the neuron id.

get_full_neuron_

get_latest_reward_event
Provides information about the last reward event.

get_neuron_ids
Return the Neuron IDs of all Neurons that have caller() as their controller or as one of their hot keys

get_neuron_ids_

get_neuron_info
Returns the public neuron info corresponding to the neuron id.

get_neuron_info_

get_pending_proposals

get_pending_proposals_

get_proposal_info

get_proposal_info_

governance
Returns an immutable reference to the global state.

governance_mut
Returns a mutable reference to the global state.

list_neurons

list_neurons_

list_neurons_pb

list_proposals

list_proposals_

list_proposals_pb

main

manage_neuron

manage_neuron_

manage_neuron_pb

neuron_stake_transfer_notification

neuron_stake_transfer_notification_pb

submit_proposal
DEPRECATED: Only callable from the registry. Use manage_neuron instead.

submit_proposal_

transfer_gtc_neuron

transfer_gtc_neuron_

update_authz

update_authz_

vote
DEPRECATED: Use manage_neuron directly instead.